feat: implement RBAC system with user, role, and permission management controllers and seeders

+++ app/Http/Controllers/Admin/UserController.php
@@ -28,16 +28,16 @@ class UserController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'name'     => 'required|string|max:255',
-            'email'    => 'required|email|unique:users,email',
+            'name' => 'required|string|max:255',
+            'email' => 'required|email|unique:users,email',
             'password' => 'required|min:8|confirmed',
-            'roles'    => 'required|array',
-            'avatar'   => 'nullable|image|mimes:jpeg,png,jpg|max:2048', // Validate the image
+            'roles' => 'required|array',
+            'avatar' => 'nullable|image|mimes:jpeg,png,jpg|max:2048', // Validate the image
         ]);
 
         $user = User::create([
-            'name'     => $request->name,
-            'email'    => $request->email,
+            'name' => $request->name,
+            'email' => $request->email,
             'password' => Hash::make($request->password),
         ]);
 
@@ -59,8 +59,8 @@ class UserController extends Controller
     public function update(Request $request, User $user)
     {
         $request->validate([
-            'name'   => 'required|string|max:255',
-            'email'  => 'required|email|unique:users,email,' . $user->id,
+            'name' => 'required|string|max:255',
+            'email' => 'required|email|unique:users,email,' . $user->id,
             'avatar' => 'nullable|image|mimes:jpeg,png,jpg|max:2048',
         ]);
 
@@ -82,6 +82,11 @@ class UserController extends Controller
         if (auth()->id() === $user->id) {
             return back()->with('error', 'You cannot delete your own account.');
         }
+        // cann't delete users with role admin
+        if ($user->hasRole('admin')) {
+            return back()->with('error', 'You cannot delete user with role admin.');
+        }
+
 
         $user->delete();
         return redirect()->route('admin.users.index')->with('success', 'User deleted.');
